+++ packages/react-instantsearch/src/components/Panel.js
@@ -20,17 +20,20 @@ class Panel extends Component {
 
   constructor(props) {
     super(props);
-    this.state = { canRefine: true };
-    this.canRefine = canRefine => {
-      this.setState({ canRefine });
+
+    this.state = {
+      canRefine: true,
     };
   }
+
+  canRefine = canRefine => {
+    this.setState({ canRefine });
+  };
+
   render() {
     return (
       <div {...cx('root', !this.state.canRefine && 'noRefinement')}>
-        <h5 {...cx('title')}>
-          {this.props.title}
-        </h5>
+        <h5 {...cx('title')}>{this.props.title}</h5>
         {this.props.children}
       </div>
     );