+++ packages/react-instantsearch/src/components/RangeInput.test.js
@@ -3,9 +3,11 @@ import PropTypes from 'prop-types';
 
 import React from 'react';
 import renderer from 'react-test-renderer';
-import { mount } from 'enzyme';
+import Enzyme, { mount, shallow } from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
+Enzyme.configure({ adapter: new Adapter() });
 
-import RangeInput from './RangeInput';
+import RangeInput, { RawRangeInput } from './RangeInput';
 
 describe('RangeInput', () => {
   it('supports passing max/min values', () => {
@@ -44,6 +46,123 @@ describe('RangeInput', () => {
     expect(tree).toMatchSnapshot();
   });
 
+  it('expect to applies changes when props have changed ', () => {
+    const wrapper = shallow(
+      <RawRangeInput
+        createURL={() => '#'}
+        refine={() => {}}
+        min={0}
+        max={100}
+        currentRefinement={{ min: 0, max: 100 }}
+        canRefine={false}
+        translate={x => x}
+      />
+    );
+
+    wrapper.setProps({
+      canRefine: true,
+      currentRefinement: {
+        min: 10,
+        max: 90,
+      },
+    });
+
+    wrapper.update();
+
+    expect(wrapper.state()).toEqual({
+      from: 10,
+      to: 90,
+    });
+  });
+
+  it("expect to don't applies changes when props don't have changed", () => {
+    const wrapper = shallow(
+      <RawRangeInput
+        createURL={() => '#'}
+        refine={() => {}}
+        min={0}
+        max={100}
+        currentRefinement={{ min: 0, max: 100 }}
+        canRefine={true}
+        translate={x => x}
+      />
+    );
+
+    wrapper.setState({
+      from: 10,
+      to: 90,
+    });
+
+    wrapper.setProps({
+      canRefine: true,
+      currentRefinement: {
+        min: 0,
+        max: 100,
+      },
+    });
+
+    wrapper.update();
+
+    expect(wrapper.state()).toEqual({
+      from: 10,
+      to: 90,
+    });
+  });
+
+  it('expect to call context canRefine when props changed', () => {
+    const context = {
+      canRefine: jest.fn(),
+    };
+
+    const wrapper = shallow(
+      <RawRangeInput
+        createURL={() => '#'}
+        refine={() => {}}
+        min={0}
+        max={100}
+        currentRefinement={{ min: 0, max: 100 }}
+        canRefine={true}
+        translate={x => x}
+      />,
+      {
+        context,
+      }
+    );
+
+    wrapper.setProps({
+      canRefine: false,
+    });
+
+    expect(context.canRefine).toHaveBeenCalledTimes(2);
+  });
+
+  it("expect to not call context canRefine when props don't have changed", () => {
+    const context = {
+      canRefine: jest.fn(),
+    };
+
+    const wrapper = shallow(
+      <RawRangeInput
+        createURL={() => '#'}
+        refine={() => {}}
+        min={0}
+        max={100}
+        currentRefinement={{ min: 0, max: 100 }}
+        canRefine={true}
+        translate={x => x}
+      />,
+      {
+        context,
+      }
+    );
+
+    wrapper.setProps({
+      canRefine: true,
+    });
+
+    expect(context.canRefine).toHaveBeenCalledTimes(1);
+  });
+
   it('refines its value on change', () => {
     const refine = jest.fn();
     const wrapper = mount(
@@ -67,8 +186,14 @@ describe('RangeInput', () => {
 
     refine.mockClear();
 
-    wrapper.find('input').first().simulate('change', { target: { value: 89 } });
-    wrapper.find('input').last().simulate('change', { target: { value: 99 } });
+    wrapper
+      .find('input')
+      .first()
+      .simulate('change', { target: { value: 89 } });
+    wrapper
+      .find('input')
+      .last()
+      .simulate('change', { target: { value: 99 } });
 
     wrapper
       .find('.ais-RangeInput__root')
@@ -95,8 +220,14 @@ describe('RangeInput', () => {
 
     const formChildren = wrapper.find('.ais-RangeInput__submit');
 
-    wrapper.find('input').first().simulate('change', { target: { value: '' } });
-    wrapper.find('input').last().simulate('change', { target: { value: '' } });
+    wrapper
+      .find('input')
+      .first()
+      .simulate('change', { target: { value: '' } });
+    wrapper
+      .find('input')
+      .last()
+      .simulate('change', { target: { value: '' } });
 
     wrapper
       .find('.ais-RangeInput__root')