+++ packages/react-instantsearch/src/connectors/connectRefinementList.js
@@ -97,7 +97,9 @@ export default createConnector({
     showMore: PropTypes.bool,
     limitMin: PropTypes.number,
     limitMax: PropTypes.number,
-    defaultRefinement: PropTypes.arrayOf(PropTypes.string),
+    defaultRefinement: PropTypes.arrayOf(
+      PropTypes.oneOfType([PropTypes.string, PropTypes.number])
+    ),
     withSearchBox: PropTypes.bool,
     searchForFacetValues: PropTypes.bool, // @deprecated
     transformItems: PropTypes.func,