+++ stories/RangeInput.stories.js
@@ -13,32 +13,52 @@ const stories = storiesOf('RangeInput', module);
 stories.addDecorator(withKnobs);
 
 stories
-  .add('default', () =>
+  .add('default', () => (
     <WrapWithHits linkedStoryGroup="RangeInput">
       <RangeInput attributeName="price" />
     </WrapWithHits>
-  )
-  .add('playground', () =>
+  ))
+  .add('with default value', () => (
     <WrapWithHits linkedStoryGroup="RangeInput">
       <RangeInput
         attributeName="price"
-        min={number('max', 0)}
-        max={number('max', 300)}
-        translations={object('translations', {
-          submit: ' go',
-          separator: 'to',
-        })}
+        defaultRefinement={{ min: 50, max: 200 }}
       />
     </WrapWithHits>
-  )
-  .add('with panel', () =>
+  ))
+  .add('with min boundaries', () => (
+    <WrapWithHits linkedStoryGroup="RangeInput">
+      <RangeInput attributeName="price" min={30} />
+    </WrapWithHits>
+  ))
+  .add('with max boundaries', () => (
+    <WrapWithHits linkedStoryGroup="RangeInput">
+      <RangeInput attributeName="price" max={500} />
+    </WrapWithHits>
+  ))
+  .add('with min / max boundaries', () => (
+    <WrapWithHits linkedStoryGroup="RangeInput">
+      <RangeInput attributeName="price" min={30} max={500} />
+    </WrapWithHits>
+  ))
+  .add('with boundaries and default value', () => (
+    <WrapWithHits linkedStoryGroup="RangeInput">
+      <RangeInput
+        attributeName="price"
+        min={30}
+        max={500}
+        defaultRefinement={{ min: 50, max: 200 }}
+      />
+    </WrapWithHits>
+  ))
+  .add('with panel', () => (
     <WrapWithHits linkedStoryGroup="RangeInput">
       <Panel title="Price">
         <RangeInput attributeName="price" />
       </Panel>
     </WrapWithHits>
-  )
-  .add('with panel but no refinement', () =>
+  ))
+  .add('with panel but no refinement', () => (
     <WrapWithHits searchBox={false} linkedStoryGroup="RangeInput">
       <Panel title="Price">
         <RangeInput attributeName="price" />
@@ -47,4 +67,18 @@ stories
         </div>
       </Panel>
     </WrapWithHits>
-  );
+  ))
+  .add('playground', () => (
+    <WrapWithHits linkedStoryGroup="RangeInput">
+      <RangeInput
+        attributeName="price"
+        min={number('min', 0)}
+        max={number('max', 500)}
+        defaultRefinement={object('default value', { min: 100, max: 400 })}
+        translations={object('translations', {
+          submit: ' go',
+          separator: 'to',
+        })}
+      />
+    </WrapWithHits>
+  ));